@@ -1,10 +1,10 @@
 ﻿using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
-using Newtonsoft.Json.Linq;
 using OPS.Core;
 using System;
 using System.Net.Mail;
+using System.Text.Json;
 
 namespace OPS.API.Controllers
 {
@@ -23,15 +23,17 @@ namespace OPS.API.Controllers
 
         [HttpPost]
         [ActionName("")]
-        public async System.Threading.Tasks.Task<IActionResult> PostAsync(JObject jsonData)
+        public async System.Threading.Tasks.Task<IActionResult> PostAsync(string jsonData)
         {
             if (jsonData == null) throw new ArgumentNullException(nameof(jsonData));
 
-            var Subject = "[OPS] Contato - " + jsonData["name"];
-            var Body = jsonData["comments"].ToString();
+            var form = JsonDocument.Parse(jsonData).RootElement;
+
+            var Subject = "[OPS] Contato - " + form.GetProperty("name");
+            var Body = form.GetProperty("comments").ToString();
             var From = new MailAddress("envio@ops.net.br", "[OPS] Operação Política Supervisionada");
             var objEmailTo = new MailAddressCollection();
-            var ReplyTo = new MailAddress(jsonData["email"].ToString(), jsonData["name"].ToString());
+            var ReplyTo = new MailAddress(form.GetProperty("email").ToString(), form.GetProperty("name").ToString());
 
 #if DEBUG
             objEmailTo.Add(new MailAddress("vanderleidenir@hotmail.com", "Vanderlei Denir"));
