@@ -1,23 +1,17 @@
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO.Compression;
-using System.Linq;
-using System.Threading.Tasks;
 using AspNetCore.CacheOutput;
-using AspNetCore.CacheOutput.Extensions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.ResponseCompression;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Microsoft.Net.Http.Headers;
-using OPS.Core;
 using OPS.Core.DAO;
+using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.IO.Compression;
 
 namespace OPS.API
 {
@@ -130,9 +124,9 @@ namespace OPS.API
             {
                 app.UseDeveloperExceptionPage();
             }
-            
+
             //app.UseHttpsRedirection();
-            app.UseCacheOutput();
+            //app.UseCacheOutput();
             app.UseRouting();
 
             // Enable CORS (Cross-Origin Requests)
